Keepalive: Add initial keepalive configuration.
This commit introduces the keepalive configuration by adding
'keepalive' module and also helper and initialization functions
that will be invoked by later commits.

This commit adds new ovsdb column "keepalive" that shows the status
of the datapath threads. This is implemented for DPDK datapath and
only status of PMD threads is reported.

Signed-off-by: Bhanuprakash Bodireddy <bhanuprakash.bodireddy@intel.com>