keepalive: Retrieve PMD status periodically.
This commit implements APIs to retrieve the PMD thread status and return
the status in the below format for each PMD thread.

  Format: pmdid="status,core id,last_seen_timestamp(epoch)"
  eg:     pmd62="ALIVE,2,1503333332575"
          pmd63="GONE,3,1503333332525"

The status is periodically retrieved by keepalive thread and stored in
keepalive_stats struc which later shall be retrieved by vswitchd thread.
In case of four PMD threads the status is as below:

   "pmd62"="ALIVE,0,1503333332575"
   "pmd63"="ALIVE,1,1503333332575"
   "pmd64"="ALIVE,2,1503333332575"
   "pmd65"="ALIVE,3,1503333332575"

Signed-off-by: Bhanuprakash Bodireddy <bhanuprakash.bodireddy@intel.com>